+++ UIInferno/UICore/_executor/_Markup/Utils.ts
@@ -2,27 +2,23 @@
 /* tslint:disable */
 
 import * as Decorate from '../_Expressions/Decorate';
-import { _FocusAttrs } from 'UICore/Focus';
 import { IControl } from 'UICommon/interfaces';
 import type { Control } from 'UICore/Base';
 import {
-   Attr,
    GeneratorError,
    GeneratorStringArray,
-   IBaseAttrs,
    CommonUtils as Common,
    invisibleNodeTagName,
-   VoidTags as voidElements,
    IGeneratorConfig,
    IGeneratorDefCollection,
    TAttributes,
-   IAttributes,
    INodeAttribute,
    TObject
 } from 'UICommon/Executor';
 
 import { NumberUtils } from 'UICommon/Utils';
 import { TemplateFunction } from 'UICommon/Base';
+import { CreateTag } from './Component';
 
 /**
  * @author Тэн В.А.
@@ -34,7 +30,7 @@ interface IControlData {
 
 const invisibleNodeRegExp = new RegExp(invisibleNodeTagName, 'g');
 const invisibleNodeHTML = '<' + invisibleNodeTagName + '></' + invisibleNodeTagName + '>';
-
+export { invisibleNodeHTML };
 // joinAttrs = Attr.joinAttrs;
 
 /**
@@ -65,7 +61,7 @@ export function asyncRenderErrorTag(inst: TObject, createTag?: Function): string
          {}
       );
    }
-   const createTagFn = createTag ? createTag : createTagDefault;
+   const createTagFn = createTag ? createTag : new CreateTag().create;
    return createTagFn('div', { attributes: decoratorObject }, []);
 }
 
@@ -77,121 +73,6 @@ export function asyncRenderErrorTag(inst: TObject, createTag?: Function): string
 export function invisibleNodeCompat(markup: string): string {
    return markup && markup.indexOf && markup.indexOf(invisibleNodeHTML) === 0 ?
       markup.replace(invisibleNodeRegExp, 'div') : markup;
-}
-
-/**
- *
- * @param tag
- * @param attrs Собственные атрибуты
- * @param children
- * @param attrToDecorate
- * @param defCollection
- * @param control
- * @returns {string}
- */
-export function createTagDefault(tag: string,
-                                 attrs: IBaseAttrs | {attributes: unknown},
-                                 children: GeneratorStringArray,
-                                 attrToDecorate?: TAttributes,
-                                 defCollection?: IGeneratorDefCollection,
-                                 control?: Control): string {
-   if (tag === invisibleNodeTagName) {
-      return invisibleNodeHTML;
-   }
-
-   if (!attrToDecorate) {
-      attrToDecorate = {};
-   }
-   if (!attrs) {
-      attrs = {attributes: {}};
-   }
-
-   let mergedAttrs = Attr.processMergeAttributes(
-       attrToDecorate.attributes as IAttributes,
-       attrs.attributes as IAttributes
-   );
-
-   _FocusAttrs.prepareTabindex(mergedAttrs);
-   // remove focus attributes from object
-   if (Common.isCompat()) {
-      // ! не вырезаем фокусные атрибуты, для совместимости. чтобы старые компоненты могли работать в новом окружении
-      // textMarkupGenerator.cutFocusAttributes(mergedAttrs);
-   } else {
-      cutFocusAttributes(mergedAttrs);
-   }
-
-   Object.keys(mergedAttrs).forEach((attrName) => {
-      if (attrName.indexOf('top:') === 0) {
-         const newAttrName = attrName.replace('top:', '');
-         mergedAttrs[newAttrName] = mergedAttrs[newAttrName] || mergedAttrs[attrName];
-         delete mergedAttrs[attrName];
-      }
-   });
-   const mergedAttrsStr = mergedAttrs
-      ? decorateAttrs(mergedAttrs, {})
-      : '';
-   // tslint:disable-next-line:no-bitwise
-   if (~voidElements.indexOf(tag)) {
-      return '<' + tag + mergedAttrsStr + ' />';
-   }
-   return '<' + tag + mergedAttrsStr + '>' + joinElements(children, undefined, defCollection) + '</' + tag + '>';
-}
-
-/**
- *
- * @param attr1
- * @param attr2
- * @returns {string}
- */
-export function decorateAttrs(attr1: TAttributes, attr2: TAttributes): string {
-   function wrapUndef(value: string): string {
-      if (value === undefined || value === null) {
-         return '';
-      } else {
-         return value;
-      }
-   }
-
-   const attrToStr = (attrs: Array<string>): string => {
-      let str = '';
-      for (const attr in attrs) {
-         if (attrs.hasOwnProperty(attr)) {
-            str += (wrapUndef(attrs[attr]) !== '' ? ' ' + (attr + '="' + attrs[attr] + '"') : '');
-         }
-      }
-      return str;
-   };
-   return attrToStr(Attr.joinAttrs(attr1, attr2));
-}
-
-const focusAttrs = [
-   'ws-creates-context',
-   'ws-delegates-tabfocus',
-   'ws-tab-cycling',
-   'ws-no-focus',
-   'attr:ws-creates-context',
-   'attr:ws-delegates-tabfocus',
-   'attr:ws-tab-cycling',
-   'attr:ws-no-focus'
-];
-
-/**
- * Скрывает атрибуты необходимые для работы системы фокусов
- * @param attributes
- * @param fn
- * @param node
- * @return {object}
- */
-export function cutFocusAttributes(attributes: TAttributes, fn?: Function, node?: HTMLElement): void {
-   focusAttrs.forEach((focusAttr: string): void => {
-      if (attributes.hasOwnProperty(focusAttr)) {
-         fn && fn(focusAttr, attributes[focusAttr]);
-         delete attributes[focusAttr];
-         if (node) {
-            node.removeAttribute(focusAttr);
-         }
-      }
-   });
 }
 
 /**
